+++ AWSSunflower/Program.fs
@@ -31,25 +31,6 @@ module Main =
                 triggers = [ EffectTrigger.AfterInit ]
             )
 
-            // Toast auto-dismiss effect
-            ctx.useEffect(
-                handler = (fun _ ->
-                    if writableModel.Current.Toasts.Length > 0 then
-                        let timer = DispatcherTimer()
-                        timer.Interval <- TimeSpan.FromSeconds 5.0
-                        timer.Tick.Add(fun _ ->
-                            if writableModel.Current.Toasts.Length > 0 then
-                                dispatch (ApiExplorer.DismissToast writableModel.Current.Toasts.Head.Id)
-                                if writableModel.Current.Toasts.Length <= 1 then timer.Stop()
-                        )
-                        timer.Start()
-                        { new IDisposable with member _.Dispose() = timer.Stop() }
-                    else
-                        { new IDisposable with member _.Dispose() = () }
-                ),
-                triggers = [ EffectTrigger.AfterChange writableModel ]
-            )
-
             // Polling + loco detection timers
             ctx.useEffect(
                 handler = (fun _ ->
@@ -73,21 +54,7 @@ module Main =
                 triggers = [ EffectTrigger.AfterInit ]
             )
 
-            TabControl.create [
-                TabControl.selectedIndex model.ActiveTab
-                TabControl.onSelectedIndexChanged (fun idx -> dispatch (ApiExplorer.SetActiveTab idx))
-                TabControl.tabStripPlacement Dock.Top
-                TabControl.viewItems [
-                    TabItem.create [
-                        TabItem.header "Serial Port"
-                        TabItem.content (ApiExplorer.serialPortTabView model dispatch)
-                    ]
-                    TabItem.create [
-                        TabItem.header "API Explorer"
-                        TabItem.content (ApiExplorer.apiExplorerTabView model dispatch)
-                    ]
-                ]
-            ]
+            ApiExplorer.mainView model dispatch
         )
 
 type MainWindow() =